Adding tabindex of -1 and changing focus of the headings based on the displayed card so that screen reader experience is not diminished
Previously: I noticed that while VoiceOver reads the success message, the black outline was completely missing. It was present in the sign-up form but not in the success message. Moreover, when I pressed Ctrl + Option + Space, I thought the dismiss button was clicked, but actually, the subscribe button was being clicked